requirements.txt
websocket-client
requests
redis
